requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
